backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
confluent-kafka==2.13.0
dj-database-url==3.1.0
Django==6.0.1
django-channels==0.7.0
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
docutils==0.22.4
idna==3.11
msgpack==1.1.2
oauthlib==3.3.1
python-dotenv==1.2.1
python-geohash==0.8.5
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
